typescript: Convert User model
