+++ server/models/PayoutMethod.ts
@@ -16,6 +16,7 @@ import sequelize, { Op } from '../lib/sequelize';
 import { objHasOnlyKeys } from '../lib/utils';
 import { RecipientAccount as BankAccountPayoutMethodData } from '../types/transferwise';
 
+import User from './User';
 import models from '.';
 
 /**
@@ -112,7 +113,7 @@ export class PayoutMethod extends Model<InferAttributes<PayoutMethod>, InferCrea
    */
   static async createFromData(
     payoutMethodData: Record<string, unknown>,
-    user: typeof models.User,
+    user: User,
     collective: typeof models.Collective,
     dbTransaction: Transaction | null,
   ): Promise<PayoutMethod> {
@@ -135,7 +136,7 @@ export class PayoutMethod extends Model<InferAttributes<PayoutMethod>, InferCrea
    */
   static async getOrCreateFromData(
     payoutMethodData: Record<string, unknown>,
-    user: typeof models.User,
+    user: User,
     collective: typeof models.Collective,
     dbTransaction: Transaction | null,
   ): Promise<PayoutMethod> {
