+++ server/models/UserToken.ts
@@ -1,8 +1,9 @@
-import type { CreationOptional, InferAttributes, InferCreationAttributes, NonAttribute } from 'sequelize';
+import type { CreationOptional, ForeignKey, InferAttributes, InferCreationAttributes, NonAttribute } from 'sequelize';
 
 import oAuthScopes from '../constants/oauth-scopes';
 import sequelize, { DataTypes, Model } from '../lib/sequelize';
 
+import User from './User';
 import models from '.';
 
 export enum TokenType {
@@ -16,14 +17,14 @@ class UserToken extends Model<InferAttributes<UserToken>, InferCreationAttribute
   public declare refreshToken: string;
   public declare refreshTokenExpiresAt?: Date;
   public declare ApplicationId: number;
-  public declare UserId: number;
+  public declare UserId: ForeignKey<User['id']>;
   public declare data: Record<string, unknown>;
   public declare scope: string[];
   public declare createdAt: CreationOptional<Date>;
   public declare updatedAt: CreationOptional<Date>;
   public declare deletedAt: CreationOptional<Date>;
 
-  public declare user?: NonAttribute<typeof models.User>;
+  public declare user?: NonAttribute<User>;
   public declare client?: NonAttribute<typeof models.Application>;
 
   hasScope(scope): boolean {
