+++ test/server/lib/guest-accounts.test.ts
@@ -25,8 +25,8 @@ describe('server/lib/guest-accounts.ts', () => {
       const user = await models.User.findOne({ where: { CollectiveId: collective.id } });
       expect(collective).to.exist;
       expect(user.email).to.eq(email);
-      expect(user.data.creationRequest.ip).to.eq('1.2.3.4');
-      expect(user.data.creationRequest.userAgent).to.eq('TestUserAgent');
+      expect(user.data.creationRequest['ip']).to.eq('1.2.3.4');
+      expect(user.data.creationRequest['userAgent']).to.eq('TestUserAgent');
     });
 
     it('Works even if a verified account already exists for this email, but does not update the profile', async () => {
